parsing done + set rfc + Makefile

+++ cube3d/project/libft/ft_atoi.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dsy <marvin@42.fr>                         +#+  +:+       +#+        */
+/*   By: othabchi <othabchi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/18 15:49:45 by dsy               #+#    #+#             */
-/*   Updated: 2019/10/20 16:11:19 by dsy              ###   ########.fr       */
+/*   Updated: 2020/06/11 19:23:17 by othabchi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int i;
 	int neg;
-	int res;
+	long res;
 
 	i = 0;
 	res = 0;
